Simple chest
Chest has a simple script on it with some interfaces.
It can be locked or open. It has an Interactable interface in order to easily interact with it. The Interact function checks if there are still locks remaining, if the player has a key, unlock one of the locks. If no locks are on the chest, open it. Also has simple opening animation